+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2018/01/21
 - [CREATE][COOKBOOK] denebola
 - [UPDATE][COOKBOOK] alterf to 1.5.0
+- [CREATE][COOKBOOK] zosma
 
 # 2017/12/31
 - [CREATE][COOKBOOK] redis